pyproject.toml
[project]
name = "fastdemark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "fire>=0.7.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
]
